advanced-api-project/requirements.txt
Django>=4.2
djangorestframework>=3.13
djangorestframework-authtoken
django-filter
